Event update

+++ app/Views/admin/events/edit-event-organizer.php
@@ -96,7 +96,7 @@
                                     <td><?php $emp = $employee_model->get($value['upload_by']); echo $emp['first_name']." ".$emp['middle_name']." ".$emp['last_name'] ?></td>
                                     <td>
                                         <a href="<?= base_url() ?>admin/edit-event-organizer/<?= $value['id'] ?>" class="btn btn-sm btn-primary">Edit</a>
-                                        <a href="<?= base_url() ?>admin/delete-event-organizer/<?= $value['id'] ?>" class="btn btn-sm btn-danger">Delete</a>
+                                        <a href="<?= base_url() ?>admin/delete-event-organizer/<?= $value['id'] ?>" class="btn btn-sm btn-danger" onclick="return confirm('Are you sure...!')">Delete</a>
                                     </td>
                                 </tr>
                             <?php } ?>
